implement a chaos actor. (+)
The chaos actor is implements a variety of illegal behaviours that
trigger violations of VM invariants. These behaviours are not found in
production code, but are important to test that the VM constraints are
properly enforced.

The chaos actor is being incubated and its behaviour and ABI be standardised
shortly. Its CID is ChaosActorCodeCID, and its singleton address is 97 (Address).
It cannot be instantiated via the init actor, and its constructor panics.

Test vectors relying on the chaos actor being deployed will carry selector
"chaos_actor=true".