introduce hints in the schema and modes in generator.
Hints are arbitrary flags that convey information to the driver.
Use hints to express facts like this vector is knowingly incorrect
(e.g. when the reference implementation is broken), or that drivers
should negate the postconditions (i.e. test that they are NOT the ones
expressed in the vector), etc.

Modes tune certain elements of how the generation and assertion of
this test vector will be conducted, such as being lenient to assertion
failures when a vector is knowingly incorrect.