feat: remove previously broken vectors (#137)
If a vector is marked as broken it'll be generated with a prefix "x--". When it gets fixed, a new vector file will be created without a prefix. This PR cleans up the old broken vector and vice versa i.e. it cleans up the working vector when a vector becomes broken.