feat: include filter match on group name (#144)
I'm often working on a group of vectors so it's useful for me to run just that group instead of just a single vector. This PR updates the include filter to also match against the vector group name, so you can include a whole group, or just a single vector.