support protocol versions via test vector variants.
This commit introduces the concept of test vector _variants_.

A single test vector can have many variants, each of which is to
be executed at a different height and with a different network version.
Variants are captured in the `preconditions/variants` array.

The builder DSL has been adapted to abstract the network version,
state tree versions, and actors versions as much as possible.

All generative suites have been reviewed and made to work with with the
refactor.

By default, all vectors are generated for all *known protocol versions*.
Known protocol versions are captured in the KnownProtocolVersions table.

If the behaviour encoded in a vector changes across protocol versions, a
VectorDef can restrict the protocol versions it wants to be generated
against. There are KnownProtocolVersions{Before,From,Between} helper
functions that query the KnownProtocolVersions table to extract relevant
protocol version ranges.

Generated vectors are now suffixed with the minimum/first protocol version
codename they are known to work with. This is to aid browsability, but file
naming may change anytime. For that reason, we encode the same data inside
the selector "min_protocol_version". This enables implementations to
deterministically and purely filter out vectors for versions they do not
support yet.

The Generator automatically merges identical variants into a single
multi-variant vector. It does so by hashing the vector output after
removing the variable fields.

The epoch field under apply_{messages,tipsets} entries is now called
`epoch_offset`. It now expresses the offset **from the variant's epoch**
at which the application is performed. This enables apply blocks to remain
unchanged across variants, despite being applied at different effective
heights. We would not be able to deduplicate variants otherwise.